requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pyhd8ed1ab_0
addict=2.4.0=pypi_0
albumentations=0.5.1=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
attr=0.3.1=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
charset-normalizer=3.0.1=pypi_0
contourpy=1.0.7=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
easydict=1.9=py_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
flake8=6.0.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.42.0=py38hce63b2e_0
idna=3.4=pypi_0
imageio=2.26.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.0.0=pyha770c72_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.14.0=py38h7f3c49e_0
ipython=8.12.3=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.12.0=py38h578d9bd_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.15.8=h780b84a_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.5=pypi_0
matplotlib=3.3.4=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pypi_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mmcv=1.4.1=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
ninja=1.11.1=pypi_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
opencv-contrib-python=4.4.0.46=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openslide-python=1.3.1=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_0
pip=22.3.1=py38h06a4308_0
pipreqs=0.5.0=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
prompt-toolkit=3.0.38=pyha770c72_0
protobuf=3.15.8=py38h709712a_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.7=pypi_0
pycodestyle=2.10.0=pypi_0
pycparser=2.21=pypi_0
pyflakes=3.0.1=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python-graphviz=0.20.1=pypi_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py38ha71036d_2
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.9.1=pypi_0
segment-anything=1.0=dev_0
setuptools=65.6.3=py38h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.5=pypi_0
sqlite=3.40.1=h5082296_0
stack_data=0.6.2=pyhd8ed1ab_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.2.28=pypi_0
timm=0.3.2=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tornado=6.1=py38h0a891b7_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=pypi_0
visdom=0.2.4=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
yapf=0.32.0=pypi_0
yarg=0.1.9=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
